refactor(poesy-question): change created_time field type to Date
Update the created_time field in the QuestionPO class from using the custom Data class
to the standard Java Date class to improve consistency and interoperability with other
parts of the system.

BREAKING CHANGE: This change may affect any code that relies on the previous type of the
created_time field (Data). Adjustments will be required where the field is used or
serialized/deserialized.